Update Adguard instructions for v4.0
Starting with v4.0 of the Adguard add-on, Adguard is no longer listening for DNS requests on `172.30.32.1`. You must now refer to the LAN address of HomeAssistant instead. 

I also added an explainer note for users upgrading a previously working configuration. I'm not certain if this is the right place for that but thought it should be published somewhere. Perhaps that should just be published to the changelog of this addon?