print qrcode as text in logs
in order to get started quickly with the wireguard addon, it would be nice to have the qr code ready in the logs.

else you'll need to use the File Editor add on, which does not allow by default access to the /ssl/wireguard folder. so this change would make it more beginner friendly.